🚚 rename: SocialButton, ProfileImageUrl 경로 수정 (#52)
